requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.6.0
streamlit>=1.32.0
pandas>=2.2.0
openpyxl>=3.1.2 